# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "narrator",
    "description": "Self-hosted TTS book reader. One global session: the server holds one book, one chapter and one playhead across every client, which is why anything cacheable is scoped by ?book= and positions can be written by name. No auth, no TLS, wildcard CORS - it assumes a private machine reached over Tailscale.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/chapter/{ci}": {
      "get": {
        "tags": [
          "session"
        ],
        "summary": "One chapter's words.",
        "description": "**`?book=` is honoured**, which the python server does not do: it answers for\nwhichever book the process last loaded and ignores the query. That one\nasymmetry is why the reader's first paint was coupled to `/api/load` at all —\nthe cheapest possible \"give me the words of chapter 576\" could not be asked\nuntil the server had been told which book it was on. Here a key that is not\nthe loaded book is served straight out of that book's text bundle, with no\nsession involved. Omitted, it still means the loaded book.",
        "operationId": "chapter",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "description": "chapter index",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "book",
            "in": "query",
            "description": "The cache key. Omitted, it means the loaded book.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterText"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chapters/render": {
      "post": {
        "tags": [
          "chapters"
        ],
        "summary": "Queue chapters for rendering, in the order given. `pack: true` also packs\neach one as it completes — that is what \"download offline\" asks for first.",
        "operationId": "chapters_render",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterSetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderResult"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "no such book — the name given is neither the loaded book nor one the library knows. A *known* book that is not the loaded one is acted on rather than refused: the order goes to that book and the worker picks it up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{ci}.json": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "chunk index -> start second for the packed chapter. Without this the m4a is\none opaque blob and every reading position in the book is lost.",
        "operationId": "chapter_manifest",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "book",
            "in": "query",
            "description": "The cache key. Omitted, it means the loaded book.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manifest"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "chapters"
        ],
        "summary": "chunk index -> start second for the packed chapter. Without this the m4a is\none opaque blob and every reading position in the book is lost.",
        "operationId": "chapter_manifest",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "book",
            "in": "query",
            "description": "The cache key. Omitted, it means the loaded book.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manifest"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{ci}.m3u8": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "An HLS playlist for the packed chapter, built on first request.",
        "description": "This is the streaming path: iOS plays it natively in a plain `<audio>`, so the\nOS owns buffering and segment retry — a blip re-fetches one six-second segment\ninstead of killing the whole load. The m4a stays the download format, because\nnative HLS bypasses the service worker and can never be the offline one.",
        "operationId": "chapter_hls",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "book",
            "in": "query",
            "description": "The cache key. Omitted, it means the loaded book.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/vnd.apple.mpegurl": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "chapters"
        ],
        "summary": "An HLS playlist for the packed chapter, built on first request.",
        "description": "This is the streaming path: iOS plays it natively in a plain `<audio>`, so the\nOS owns buffering and segment retry — a blip re-fetches one six-second segment\ninstead of killing the whole load. The m4a stays the download format, because\nnative HLS bypasses the service worker and can never be the offline one.",
        "operationId": "chapter_hls",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "book",
            "in": "query",
            "description": "The cache key. Omitted, it means the loaded book.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/vnd.apple.mpegurl": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{ci}.m4a": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "The packed chapter: AAC-LC mono 64 kbit/s, 24 kHz, `+faststart`. **Serves\nbyte ranges** — iOS probes with a Range before it will play anything. This is\nthe *download* format.",
        "operationId": "chapter_audio",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "book",
            "in": "query",
            "description": "The cache key. Omitted, it means the loaded book.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "audio/mp4": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "206": {
            "description": "byte range"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "416": {
            "description": "unsatisfiable range"
          }
        }
      },
      "head": {
        "tags": [
          "chapters"
        ],
        "summary": "The packed chapter: AAC-LC mono 64 kbit/s, 24 kHz, `+faststart`. **Serves\nbyte ranges** — iOS probes with a Range before it will play anything. This is\nthe *download* format.",
        "operationId": "chapter_audio",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "book",
            "in": "query",
            "description": "The cache key. Omitted, it means the loaded book.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "audio/mp4": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "206": {
            "description": "byte range"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "416": {
            "description": "unsatisfiable range"
          }
        }
      }
    },
    "/api/chunk/{ci}/{i}.wav": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "One rendered chunk. GET or HEAD; **404 means \"not rendered yet\", not an\nerror** — the reader polls this while the renderer catches up.",
        "operationId": "chunk_wav",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "i",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "404": {
            "description": "not rendered yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "chapters"
        ],
        "summary": "One rendered chunk. GET or HEAD; **404 means \"not rendered yet\", not an\nerror** — the reader polls this while the renderer catches up.",
        "operationId": "chunk_wav",
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "i",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "404": {
            "description": "not rendered yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hls/{book}/{ci}/{name}": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "One fMP4 segment (or the init segment) of a chapter's HLS rendition. `name`\nis validated against `init.mp4|segNNNNN.m4s`; anything else is a 404.",
        "operationId": "hls_segment",
        "parameters": [
          {
            "name": "book",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "video/mp4": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "chapters"
        ],
        "summary": "One fMP4 segment (or the init segment) of a chapter's HLS rendition. `name`\nis validated against `init.mp4|segNNNNN.m4s`; anything else is a 404.",
        "operationId": "hls_segment",
        "parameters": [
          {
            "name": "book",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "video/mp4": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/position": {
      "post": {
        "tags": [
          "vault"
        ],
        "summary": "Record a position for a *named* book without loading it.",
        "description": "This is what the offline queue pushes on reconnect, possibly for a book the\nserver has since swapped out. Last write wins, which is the right rule for one\nreader on several devices. Fields not supplied are kept from the existing\nrecord.",
        "operationId": "position",
        "parameters": [
          {
            "name": "X-Narrator-Device",
            "in": "header",
            "description": "Who is reporting. Additive: omitted, this is the anonymous legacy device and the behaviour is exactly the pre-device one. Supplied, it is echoed on the `position` event so other devices can tell a real move from their own echo.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Narrator-Device-Name",
            "in": "header",
            "description": "A human label for that device, for display only.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok2"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/renderer": {
      "post": {
        "tags": [
          "session"
        ],
        "summary": "Explicitly stop/start the render worker, independent of playback.",
        "operationId": "renderer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RendererBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RendererResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Every error body in this API: `{\"error\": \"...\"}`, and nothing else. The\nreader's `get()` reads exactly this field.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ChapterSetBody": {
        "type": "object",
        "properties": {
          "book": {
            "type": [
              "string",
              "null"
            ],
            "description": "The book these chapters belong to. **Additive, and the one worth\nsending**: this is the endpoint where a race costs hours. The reader\npolls `/api/chapters`, the reader taps \"download the rest\", and in\nbetween the watcher may have picked up an epub, the Obsidian plugin may\nhave opened something, another device may have loaded another book — and\n74 chapters of rendering then land on that one. Omitted, it means\n\"whatever is loaded\", as before.\n\nSupplied and naming a *different* book that the library knows, the order\nis placed on **that** book — the worker follows standing orders across\nthe library and the packer packs them, so naming one is a request the\nserver can honour rather than a race it has to refuse. Supplied and\nnaming nothing the library has, it is a **409**, which is now a real\nrefusal rather than a limitation wearing one's clothes."
          },
          "chapters": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "force": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "`/api/chapters/build` only: rebuild even if an m4a already exists."
          },
          "pack": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "`/api/chapters/render` only: also pack each chapter as it completes."
          }
        }
      },
      "ChapterText": {
        "type": "object",
        "required": [
          "i",
          "title",
          "chunks",
          "paras"
        ],
        "properties": {
          "chunks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "i": {
            "type": "integer",
            "minimum": 0
          },
          "paras": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 0
            },
            "description": "What lets the reader re-form real paragraphs from chunks. Additive:\nexisting clients (the Obsidian plugin) only read title + chunks."
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "Ok2": {
        "type": "object",
        "description": "`{\"ok\": true}` plus whatever else a handler wants to say.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "RendererBody": {
        "type": "object",
        "properties": {
          "on": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}